+++ src/main/java/org/openjfx/ProjectItemController.java
@@ -4,6 +4,7 @@ import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
 import javafx.scene.control.Button;
 import javafx.scene.control.Label;
+import models.Project;
 
 import java.net.URL;
 import java.util.ResourceBundle;
